Refactoring uploader into parent-child class structure
