+++ trunk-recorder/uploaders/openmhz_uploader.h
@@ -1,0 +1,14 @@
+#ifndef OPENMHZ_UPLOADER_H
+#define OPENMHZ_UPLOADER_H
+
+
+class Uploader;
+
+#include "uploader.h"
+
+class OpenmhzUploader : public Uploader {
+public:
+    int upload(struct call_data_t *call);
+};
+
+#endif
